Changed the generic_output_int function to accept a uint64_t, to prevent outputting -ve ints when above max int (e.g. outputting a large number of timesteps).

+++ src/vio/outputfunctions.cpp
@@ -27,7 +27,7 @@
 
 namespace vout{
 	// Output Function 0
-    std::string generic_output_int(std::string str,int i, bool header){
+    std::string generic_output_int(std::string str,uint64_t i, bool header){
       std::ostringstream res;
       vout::fixed_width_output result(res,vout::fw_size_int); 
       if(header){
